+++ resources/lang/en/people.php
@@ -114,4 +114,6 @@ return [
     'todays_transactions' => 'Todays transactions',
     'user_added_coupons_from_project' => '<strong>:user</strong> added <strong>:amount :coupons</strong> from <strong>:project</strong> for <strong>:date</strong>',
     'user_removed_coupons_from_project' => '<strong>:user</strong> removed <strong>:amount :coupons</strong> from <strong>:project</strong> for <strong>:date</strong>',
+    'import_people_data' => 'Import people data',
+    'import_bank_data' => 'Import bank data',
 ];